+++ engine/cmake/volkano_version.in
@@ -9,7 +9,7 @@
 
 #include <string_view>
 
-#include "engine/core/int_types.h"
+#include "core/int_types.h"
 
 namespace volkano {
