+++ engine/include/core/algo/index_of.h
@@ -10,7 +10,7 @@
 #include <ranges>
 #include <utility>
 
-#include "engine/core/int_types.h"
+#include "core/int_types.h"
 
 namespace volkano::algo {
