+++ engine/include/core/math/constants.h
@@ -10,7 +10,7 @@
 #include <concepts>
 #include <numbers>
 
-#include "engine/core/int_types.h"
+#include "core/int_types.h"
 
 namespace volkano::math::consts {
